[HUDI-121] Fix licensing issues found during RC voting by general incubator group
