+++ hudi-utilities/src/test/java/org/apache/hudi/utilities/TestHoodieDeltaStreamer.java
@@ -25,6 +25,7 @@ import static org.junit.Assert.fail;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Random;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
@@ -529,31 +530,17 @@ public class TestHoodieDeltaStreamer extends UtilitiesTestBase {
   public static class DistanceUDF implements UDF4<Double, Double, Double, Double, Double> {
 
     /**
-     * Taken from https://stackoverflow.com/questions/3694380/calculating-distance-between-two-points-using-latitude-
-     * longitude-what-am-i-doi Calculate distance between two points in latitude and longitude taking into account
-     * height difference. If you are not interested in height difference pass 0.0. Uses Haversine method as its base.
-     *
-     * lat1, lon1 Start point lat2, lon2 End point el1 Start altitude in meters el2 End altitude in meters
-     *
-     * @returns Distance in Meters
+     * Returns some random number as distance between the points
+     * 
+     * @param lat1 Latitiude of source
+     * @param lat2 Latitude of destination
+     * @param lon1 Longitude of source
+     * @param lon2 Longitude of destination
+     * @return
      */
     @Override
     public Double call(Double lat1, Double lat2, Double lon1, Double lon2) {
-
-      final int R = 6371; // Radius of the earth
-
-      double latDistance = Math.toRadians(lat2 - lat1);
-      double lonDistance = Math.toRadians(lon2 - lon1);
-      double a = Math.sin(latDistance / 2) * Math.sin(latDistance / 2) + Math.cos(Math.toRadians(lat1))
-          * Math.cos(Math.toRadians(lat2)) * Math.sin(lonDistance / 2) * Math.sin(lonDistance / 2);
-      double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-      double distance = R * c * 1000; // convert to meters
-
-      double height = 0;
-
-      distance = Math.pow(distance, 2) + Math.pow(height, 2);
-
-      return Math.sqrt(distance);
+      return new Random().nextDouble();
     }
   }
